Multichannel bug is fixed.

+++ Src/AutosarBCM/AutosarBCM/Common/FormHardwareList.cs
@@ -89,7 +89,7 @@ namespace AutosarBCM.Common
                 else if(cmbDevices.SelectedItem is IntrepidCsCan intrepidCsCan)
                 {
                     intrepidCsCan.BitRate = Convert.ToUInt32(string.IsNullOrWhiteSpace(grpIntrepidCanProperties_cmbBitRate.Text) ? "0" : grpIntrepidCanProperties_cmbBitRate.Text);
-                    intrepidCsCan.NetworkID = string.IsNullOrWhiteSpace(grpIntrepidCanProperties_cmbNetworkId.Text) ? (uint)CSnet.eNETWORK_ID.NETID_DEVICE : Convert.ToUInt32(Enum.Parse(typeof(CSnet.eNETWORK_ID), grpIntrepidCanProperties_cmbNetworkId.Text));
+                    intrepidCsCan.NetworkID = string.IsNullOrWhiteSpace(grpIntrepidCanProperties_cmbNetworkId.Text) ? (uint)CSnet.eNETWORK_ID.NETID_DEVICE : Convert.ToByte(Enum.Parse(typeof(CSnet.eNETWORK_ID), grpIntrepidCanProperties_cmbNetworkId.Text));
                 }
                 else if (cmbDevices.SelectedItem is KvaserCan kvaserCan)
                 {
